% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{clusterMaker}
\alias{clusterMaker}
\title{A function to}
\usage{
clusterMaker(chr, pos, assumeSorted = FALSE, maxGap = 300)
}
\arguments{
\item{assumeSorted}{Defaults to FALSE}

\item{maxGap}{Defaults to 300}
}
\value{
list of segments
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

