% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{DMRfinder}
\alias{DMRfinder}
\title{A function for the WGBS version of the 'BumphunterEngine' function with 'LM'
and 'GLM'}
\usage{
DMRfinder(meth.mat, unmeth.mat, minInSpan = 10, design, method = "LM",
  chr = NULL, pos, cluster = NULL, coef = 2, minNum = 70,
  bpSpan = 1000, maxGapSmooth = 1e+08, minNumRegion = 5,
  betabin = FALSE, cutoff = NULL, pickCutoff = FALSE,
  pickCutoffQ = 0.99, maxGap = 500, nullMethod = "GLMM", smooth = FALSE,
  smoothFunction = locfitByCluster, useWeights = TRUE,
  B = ncol(permutations), permutations = NULL, verbose = TRUE,
  workers = NULL, loci = TRUE, subject = TRUE,
  sampleSize = (ncol(meth.mat)/2))
}
\arguments{
\item{minInSpan}{Defaults to 10}

\item{method}{Defaults to LM}

\item{chr}{Defaults to NULL}

\item{cluster}{Defaults to NULL}

\item{minNum}{Defaults to 70}

\item{bpSpan}{Defaults to 1000}

\item{maxGapSmooth}{Defaults to 1e8}

\item{minNumRegion}{Defaults to 5}

\item{betabin}{Defaults to FALSE}

\item{cutoff}{Defaults to NULL}

\item{pickCutoff}{Defaults to FALSE}

\item{pickCutoffQ}{Defaults to 0.99}

\item{maxGap}{Defaults to 500}

\item{nullMethod}{Defaults to GLMM}

\item{smooth}{Defaults to FALSE}

\item{smoothFunction}{locfitByCluster}

\item{useWeights}{Defaults to TRUE}

\item{B}{Defaults to ncol(permutations)}

\item{permutations}{Defaults to NULL}

\item{verbose}{Defaults to TRUE}

\item{workers}{Defaults to NULL}

\item{loci}{Defaults to TRUE}

\item{subject}{Defaults to TRUE}

\item{sampleSize}{Defaults to ncol(meth.mat)/2}
}
\description{
This function is the WGBS version of the 'BumphunterEngine' function with 'LM'
and 'GLM'
}
\keyword{inference}

