% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{estim}
\alias{estim}
\title{A function for the WGBS version of the '.getEstimate' function in Bumphunter}
\usage{
estim(meth.mat, unmeth.mat, design, method = "LM", coef, B = NULL,
  permutations = NULL, full, workers = NULL)
}
\arguments{
\item{method}{Defaults to LM}

\item{B}{Defaults to NULL}

\item{permutations}{Defaults to NULL}

\item{workers}{Defaults to NULL}
}
\description{
This function is the WGBS version of the '.getEstimate' function in
Bumphunter for computing the estimated difference of population means for the
two biological groups
}
\keyword{inference}

