% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{getEstimate}
\alias{getEstimate}
\title{A function to compute the coefficient estimates for regression of the
methylation levels on the group indicator variable, at each CpG site}
\usage{
getEstimate(mat, design, coef, B = NULL, permutations = NULL,
  full = FALSE)
}
\arguments{
\item{B}{Defaults to NULL}

\item{permutations}{Defaults to NULL}

\item{full}{Defaults to FALSE}
}
\description{
This function computes the coefficient estimates for regression of the
methylation levels on the group indicator variable, at each CpG site
}
\keyword{inference}

