% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{getEstimatePooled}
\alias{getEstimatePooled}
\title{A function to}
\usage{
getEstimatePooled(meth.mat, unmeth.mat, design, coef)
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

