% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{getEstimateWeighted}
\alias{getEstimateWeighted}
\title{A function to}
\usage{
getEstimateWeighted(mat, wt.mat, design, coef)
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

