% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{getSegments}
\alias{getSegments}
\title{A function to}
\usage{
getSegments(x, f = NULL, cutoff = quantile(abs(x), 0.99),
  assumeSorted = FALSE, verbose = FALSE)
}
\arguments{
\item{cutoff}{Defaults to quantile(abs(x), 0.99)}

\item{assumeSorted}{Defaults to FALSE}

\item{verbose}{Defaults to FALSE}
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

