% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{locfitByCluster0}
\alias{locfitByCluster0}
\title{A function to}
\usage{
locfitByCluster0(dat = NULL, minNum = 7, bpSpan = 1000, minInSpan = 0,
  verbose = TRUE)
}
\value{
list of fitted clusters smoothened by locfit
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

