% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designOligoPool.R, R/mapToBarcodes.R
\name{mapToBarcodes}
\alias{mapToBarcodes}
\title{Design oligo pool by tiling a set of sequences}
\usage{
mapToBarcodes(fastqCases, fastqControls, conditionLabels = c("case",
  "control"), oligoMap = "oligoMap.fa", oligoOut = "oligoOut")

mapToBarcodes(fastqCases, fastqControls, conditionLabels = c("case",
  "control"), oligoMap = "oligoMap.fa", oligoOut = "oligoOut")
}
\arguments{
\item{fastqCases}{character vector with the name of 'case' fastq files.}

\item{fastqControls}{character vector with the name of 'control' fastq files.}

\item{conditionLabels}{character vector of length two which contains the 
condition labels for the two conditions that are being compared. Default 
value is c("case", "control")}

\item{oligoMap}{the name of the tab seperated file which provides a link 
between oligo name, barcode and variable sequence. The file should have 
columns for the name, barcode sequence and variable sequence of the oligo.}

\item{oligoOut}{the name of the directory where the output file will be 
written. If the output directory does not exist, one will be created. 
Defaults to oligoOut}

\item{regionsFile}{the name of the file with regions to tile for the oligo 
pool. Please make sure the file is in FASTA format and has unique headers 
for sequence of each region.}

\item{tileSize}{numeric indicating the size of each individual oligo. The 
tile size is also sometimes called the length of variable sequence. Since 
universal primer sites, barcodes and potentially a Restriction Enzyme site 
has to be added to the tile for a complete oligo, it is usually ~50 bp less
than the maximum length of the oligo you can synthesize. Default value is 
110.}

\item{overlap}{numeric describing how many nucleotides you wish to move 
between 2 successive tiles. Sometimes, this is called the step size and 
this number will determine the resolution of your data downstream. Default 
value is 10.}

\item{barcodesPerSequence}{numeric for the number of unique barcodes to use 
for each variable sequence. In certain applications, the user might wish to
use upto 20 barcodes for each individual tile for more accurate 
measurement. Default value is 1.}

\item{barcodesFile}{the name of the file with a list of barcodes you wish to 
use for designing the oligo pool. You don't need to give this file since 
the script can generate barcodes on its own but is given to the user as an 
option. If you provide the file, please note \code{microSeedsFile} and 
\code{badNucs} are redundant. Default value is NULL.}

\item{univPrimers}{character vector of the universal primer sequences added 
upstream and downstream in each oligo. Provide in the order 
c('ForwardPrimer', 'ReversePrimer'). Default value is 
c('ACTGGCCGCTTCACTG','AGATCGGAAGAGCGTCG'). Use non-standard primer 
sequences at your own risk.}

\item{reSeq}{character describing the restriction enzyme site you wish to 
add. To play some games, a restriction enzyme site is neccessary but this 
is not strictly required. Defaults to NULL.}

\item{microSeedsFile}{the name of a file with microRNA seeds. These sequences
are required to ensure they are not present in any barcodes. Please ensure 
the file is gzipped and contains one microRNA seed sequence per line. This 
is a required parameter if \code{barcodeFile} is not given. Default value 
is NULL.}

\item{badNucs}{character vector with potential sub strings you wish to avoid 
in barcodes. For example, triple N's are hard to PCR and so it is a good 
idea to avoid them in barcodes. This option is highly recommended if 
\code{barcodeFile} is not specified. Default value is c('AAA','TTT', 'CCC',
'GGG').}

\item{numScrambles}{numeric describing the number of times each tile 
(variable sequence) should be scrambled. Default is 0.}

\item{outDir}{the name of a directory you want to write the oligo pool 
sequences to. The directory will be created if not present already. Default
value is lncTilingGame}
}
\description{
This function helps a user design an oligo pool based on their needs. Many
different options provided to the user allow a high degree of customization
and flexibility. There are 3 main steps in this function: (1). Generate a
potential set of barcodes avoiding the presence of any unwanted sub strings
such as microRNA seeds (2) Tile each sequence in the input file using the
specified tile size and overlap (3) Generate final oligo pool by
concatenating the universal primers, potential restriction enzyme sequence
and a unique barcode with each tile. Each oligo is designed as
UPS_FP----VarSeq----RESite----Barcode----UPS_RP

This function allows you to map your sequencing files back to the oligo pool
and compile a table describing counts for each oligo. For a read to map to an
oligo, 2 conditions must be satisfied: (1) The first 10 nucleotides of the
read should perfectly map one barcode in the pool and (2) The remaining
nucleotides should have no more than 2 mismatches to the variable sequence
linked with the barcode.
}
\keyword{design}
\keyword{fastq}
\keyword{mapping}
\keyword{oligo}
\keyword{pool}

