% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelNucCounts.R
\name{modelNucCounts}
\alias{modelNucCounts}
\title{A function to normalize counts}
\usage{
modelNucCounts(normalizedCounts, metaData, modelMethod = c("median", "sum",
  "pgm"), oligoLen = NULL)
}
\arguments{
\item{normalizedCounts}{matrix of normalized counts. Defaults to
allTranscripts_Norm.tab}

\item{metaData}{A file with the meta data of the experiment. Should include
details such as the name of genes tiled, window size, length of oligo.
Please see oligoMeta.tab in extdata for reference.}

\item{modelMethod}{The modeling method used to get nucleotide counts from
oligo counts. Must be either "median", "sum", or "pgm".  Defaults to 
"median".}

\item{oligoLen}{numeric value for the length of the oligos if all oligos have
the same value and this information is not included in the metadata file.}
}
\value{
modeledNucs matrix of modeled nucleotide counts
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{modeling}

