% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{normalize}
\alias{normalize}
\title{A function to normalize counts}
\usage{
normalize(rawCounts = "allTranscriptsCounts_Raw.tsv", normType = "median",
  quantile = 0.5)
}
\arguments{
\item{rawCounts}{the name of the file with counts for each oligo. Contains 
one row per oligo counts with each sample its own column. Please see 
\code{system.file("inst", "extdata", "allTranscriptsCounts_Raw.tsv", package =
"oligoGames")} for an example. Defaults to allTranscriptsCounts_Raw.tsv}

\item{normType}{The method of library normalization the user prefers. Two 
options: median or quantile. Defaults to median}

\item{quantile}{The quantile the user wishes to use. Should be a numeric 
between 0 and 1.}
}
\value{
normCounts matrix of normalized counts
}
\description{
This function performs library normalization for the experiment counts
}
\examples{
rawCounts <- system.file("inst", "extdata", "allTranscriptsCounts_Raw.tsv", package = "oligoGames")
normalizedCounts <- normalize(rawCounts, normType="median")
}
\keyword{normalization}

