% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{normCounts}
\alias{normCounts}
\title{A function to normalize counts}
\usage{
normCounts(countsFile = "allTranscriptsCounts_Raw.tsv", normType = "median",
  quantile = 0.5)
}
\arguments{
\item{countsFile}{string with file path to the counts file. Defaults to
allTranscriptsCounts_Raw.tsv}

\item{normType}{The method of library normalization the user prefers. Two
options: median or quantile. Defaults to median}

\item{quantile}{The quantile the user wishs to use. Should be a number
between 0 and 1.}
}
\value{
normCounts matrix of normalized counts
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{normalization}

