% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{reduceIt}
\alias{reduceIt}
\title{A function to}
\usage{
reduceIt(x, elem, bind = rbind)
}
\arguments{
\item{bind}{Defaults to rbind}
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

