% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{regionFinder}
\alias{regionFinder}
\title{A function to}
\usage{
regionFinder(x, chr, pos, cluster = NULL, y = x, ind = seq(along = x),
  order = TRUE, minNumRegion = 5, maxGap = 300,
  cutoff = quantile(abs(x), 0.99), assumeSorted = FALSE,
  meth.mat = meth.mat, unmeth.mat = unmeth.mat, verbose = TRUE,
  design = design, workers = workers, betabin)
}
\arguments{
\item{cluster}{Defaults to NULL}

\item{y}{Defaults to x}

\item{order}{Defaults to TRUE}

\item{minNumRegion}{Defaults to 5}

\item{maxGap}{Defaults to 300}

\item{cutoff}{Defaults to quantile(abs(x), 0.99)}

\item{assumeSorted}{Defaults to FALSE}

\item{meth.mat}{Defaults to meth.mat}

\item{unmeth.mat}{Defaults to unmeth.mat}

\item{verbose}{Defaults to TRUE}

\item{design}{Defaults to design}

\item{workers}{Defaults to workers}
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

