% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRfinder.R
\name{DRfinder}
\alias{DRfinder}
\title{Perform inference to detect differential regions in oligo experiments}
\usage{
DRfinder(OligoSignal, conditionLabels = c("condition1", "condition2"),
  minInSpan = 10, bpSpan = 100, minNumRegion = 5, cutoff = NULL,
  smooth = FALSE, verbose = TRUE, workers = NULL,
  sampleSize = (ncol(OligoSignal) - 1)/2, maxPerms = 50, logT = TRUE,
  coef = 2, onlyUp = FALSE)
}
\arguments{
\item{OligoSignal}{a data frame in the format returned by the function
\code{modelNucCounts}. Contains one row per nucleotide count.
 The first column contains the basepair positions of the nucleotides and
 the remaining columns hold the counts themselves (one column per sample).}

\item{conditionLabels}{character vector of length two which contains the 
condition labels for the two conditions that are being compared.}

\item{minInSpan}{positive integer that represents the minimum number of
nucleotides in a smoothing span window if \code{smooth} is TRUE.  
Default value is 10.}

\item{bpSpan}{a positive integer that represents the length in basepairs
of the smoothing span window if \code{smooth} is TRUE.  Default value is 
100}

\item{minNumRegion}{positive integer that represents the minimum number of
nucleotides to consider for a candidate region. Default value is 5.}

\item{cutoff}{scalar value that represents the absolute value (or a vector 
of two numbers representing a lower and upper bound) for the cutoff of 
the single nucleotide condition coefficient that is used to discover 
candidate regions.}

\item{smooth}{logical value that indicates whether or not to smooth the 
nucleotide level signal when discovering candidate regions.
Defaults to FALSE.}

\item{verbose}{logical value that indicates whether progress messages
should be printed to stout. Defaults value is TRUE.}

\item{workers}{positive integer that represents the number of cores to 
use if parallelization is desired of the smoothing step.}

\item{sampleSize}{positive integer that represents the number of samples
in each condition.  Defaults to \code{(ncol(OligoSignal)-1)/2}.}

\item{maxPerms}{a positive integer that represents the maximum number 
of permutations that will be used to generate the global null 
distribution of test statistics.}

\item{logT}{logical value that indicates whether to model the log2 
transformed signal (plus a pseudocount of 1).  Default is TRUE.  Only
set to false if transformation has been done prior to running this 
function, or if distribution of raw values looks relatively symmetric.}

\item{coef}{positive integer that indicates which column of the design
matrix in \code{design} contains the condition covariate of interest}

\item{onlyUp}{a logical value indicating whether to only consider differences
in the positive direction (signal in condition 1 - condition 2 > 0). 
Default value is FALSE.}
}
\value{
a data.frame that contains the results of the inference. The
   data.frame contains one row for each candidate region, and 
   9 columns, in the following order: 1. chr = 
   region level labels such as chromosome, gene, or lncRNA, 2. start = 
   start basepair position of the region, 3. end = end basepair position
   of the region,
   4. indexStart = the index of the region's starting nucleotide, 
   5. indexEnd = the index of the region's ending nucleotide,
   6. length = the number of nucleotides contained in the region,
   7. stat = the test statistic for the condition difference,
   8. pval = the permutation p-value for the significance of the test
   statistic, and 9. qval = the q-value for the test statistic (adjustment
   for multiple comparisons to control false discovery rate).
}
\description{
This is the main inference function that aims to find regions with
differential signal between two conditions.  The two main steps of the 
procedure are (1) detect candidate regions from the nucleotide level 
signal, optionally smoothing to combat loss of power due to low coverage, 
and (2) evaluate the test statistic of condition difference across each
candidate region by comparing to a global null distribution generated 
by permuting sample labels.
}
\examples{
\dontrun{
normalizedCounts <- normCounts(rawCounts = system.file("extdata", 
"allTranscriptsCounts_Raw.tsv", package = "oligoGames"))
metaData <- system.file("extdata", "oligoMeta.tsv", package = "oligoGames")
oligoLen <- 110
conditionLabels <- c("Nuclei", "Total")
modeledNucs <- modelNucCounts(normalizedCounts, metaData, 
conditionLabels, modelMethod = "median", oligoLen = 110)
DRregions <- DRfinder(modeledNucs, conditionLabels, minInSpan = 5, 
bpSpan = 50, minNumRegion = 3, cutoff = 0.05, smooth = TRUE, verbose = TRUE, 
workers = 1, sampleSize = 4, maxPerms = 50)
}
}
\keyword{inference}

