% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToBarcodes.R
\name{mapToBarcodes}
\alias{mapToBarcodes}
\title{A function to map to barcodes}
\usage{
mapToBarcodes(fastqCases, fastqControls, labels, oligoMap, oligoOut)
}
\arguments{
\item{fastqCases}{string of comma seperated fastq files for cases}

\item{fastqControls}{string of comma seperated fastq files for controls}

\item{labels}{string of comma seperated labels for case and control. Defaults
to nuclear,total}

\item{oligoMap}{string with the path to a file linking each barcode to a
variable sequence and transcript ID. Defaults to oligoMap.tsv}

\item{oligoOut}{string with the path to the directory where all output files
will be written. If the output directory does not exist, one will be
created. Defaults to oligoOut}
}
\description{
This function allows you to map fastq files to barcodes and get counts table
}
\keyword{fastq}
\keyword{mapping}

