% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToBarcodes.R
\name{mapToBarcodes}
\alias{mapToBarcodes}
\title{Map case and control fastq files to unique oligo barcodes and generate counts
table}
\usage{
mapToBarcodes(fastqCases, fastqControls, conditionLabels = c("case",
  "control"), oligoMap = "oligoMap.fa", oligoOut = "oligoOut")
}
\arguments{
\item{fastqCases}{character vector with the name of 'case' fastq files.}

\item{fastqControls}{character vector with the name of 'control' fastq files.}

\item{conditionLabels}{character vector of length two which contains the 
condition labels for the two conditions that are being compared. Default 
value is c("case", "control")}

\item{oligoMap}{the name of the tab seperated file which provides a link 
between oligo name, barcode and variable sequence. The file should have 
columns for the name, barcode sequence and variable sequence of the oligo.}

\item{oligoOut}{the name of the directory where the output file will be 
written. If the output directory does not exist, one will be created. 
Defaults to oligoOut}
}
\description{
This function allows you to map your sequencing files back to the oligo pool
and compile a table describing counts for each oligo. For a read to map to an
oligo, 2 conditions must be satisfied: (1) The first 10 nucleotides of the
read should perfectly map one barcode in the pool and (2) The remaining
nucleotides should have no more than 2 mismatches to the variable sequence
linked with the barcode.
}
\keyword{fastq}
\keyword{mapping}

