% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{smoother}
\alias{smoother}
\title{A function to}
\usage{
smoother(y, x = NULL, weights = NULL, chr = chr,
  maxGapSmooth = maxGapSmooth, smoothFunction, verbose = TRUE,
  workers = 1, minNum = minNum, minInSpan = minInSpan, bpSpan = bpSpan)
}
\arguments{
\item{x}{Defaults to NULL}

\item{weights}{Defaults to NULL}

\item{verbose}{Defaults to TRUE}

\item{workers}{Defaults to 1}
}
\value{
A function to

This function performs library normalization for the experiment counts
}
\description{
This function performs library normalization for the experiment counts
}
\keyword{inference}

